flask/requirements.txt
Flask_Cors==3.0.8
Flask==1.0.2
python-dotenv==0.13.0
uwsgi
dnspython
kafka-python==2.0.1
Flask-SQLAlchemy==2.4.4
redis==3.5.3
boto3==1.14.60
SQLAlchemy==1.3.19
psycopg2==2.8.6